+++ jsconfig.json
@@ -8,7 +8,8 @@
             "@assets": ["./src/assets"],
             "@icons": ["./src/assets/icons"],
             "@ui": ["./src/ui"],
-            "@styles": ["./src/styles"]
+            "@styles": ["./src/styles"],
+            "@utils": ["./src/utils"]
         }
     }
 }
