+++ src/components/Homepage/ScheduleDetail.jsx
