+++ src/components/Sidebar/index.js
@@ -1,0 +1,5 @@
+export { default as GroupMenu } from "./GroupMenu";
+export { default as GroupMenuItem } from "./GroupMenuItem";
+export { default as Header } from "./Header";
+export { default as Sidebar } from "./Sidebar";
+export { default as SidebarMenu } from "./SidebarMenu";
