+++ src/utils/time.js
@@ -1,0 +1,9 @@
+export const updateClock = () => {
+    let now = new window.Date();
+    const month = (now.getMonth() + 1).toString().padStart(2, "0");
+    const day = now.getDate().toString().padStart(2, "0");
+    const hours = now.getHours().toString().padStart(2, "0");
+    const minutes = now.getMinutes().toString().padStart(2, "0");
+    const seconds = now.getSeconds().toString().padStart(2, "0");
+    return { month, day, hours, minutes, seconds };
+};
